Move eightball and misc responses to snark plugin
