+++ bot.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-import os, re, random, hashlib, datetime
+import os, re, random
 import dotenv, aiohttp
 import hikari, lightbulb
 
@@ -15,12 +15,6 @@ bot = lightbulb.BotApp(
         )
 
 
-eight_ball_responses = [ "It is certain.", "It is decidedly so.", "Without a doubt.", "Yes, definitely.",
-               "You may rely on it.", "As I see it, yes.", "Most likely.", "Outlook good.",
-               "Yes.", "Signs point to yes.", "Reply hazy, try again.", "Ask again later.",
-               "Better not tell you now.", "Cannot predict now.", "Concentrate and ask again.",
-               "Don't count on it.", "My reply is no.", "My sources say no.", "Outlook not so good.", "Very Doubtful."]
-
 @bot.listen(hikari.StartedEvent)
 async def botStartup(event):
     print("Bot has started up!")
@@ -33,37 +27,6 @@ async def botStartup(event):
 async def ping(ctx: lightbulb.Context) -> None:
     await ctx.respond(f"Pong! Latency: {bot.heartbeat_latency*1000:.2f}ms")
 
-def choose_eightball_response(message):
-    # Add current date down to hour precision to vary the response periodically
-    hash_input = message + datetime.datetime.now().strftime('%Y%m%d%H')
-    index = hashlib.md5(hash_input.encode()).digest()[0] % len(eight_ball_responses)
-    return eight_ball_responses[index]
-    
-def findWholeWord(word, text):
-    return re.compile(r'\b({0})\b'.format(word), flags=re.IGNORECASE).search(text)
-
-@bot.listen(hikari.GuildMessageCreateEvent)
-async def tags_bot(event):
-    if event.is_bot or not event.content:
-        return
-    mentioned_ids = event.message.mentions.user_ids
-    if bot.application.id not in mentioned_ids:
-        return
-    messageContent = event.content
-    regexp = re.compile('(\S|\s)\?(\s|$)')
-    if regexp.search(messageContent):
-        print("Giving 8-ball response.")
-        await event.message.respond(choose_eightball_response(messageContent))
-    elif findWholeWord('broken', messageContent):
-        await event.message.respond(f"No {event.author.mention}, you're broken :disguised_face:")
-    elif findWholeWord('thanks', messageContent) or findWholeWord('thank', messageContent):
-        await event.message.respond(f"You're welcome :heart:")
-    elif findWholeWord('work', messageContent):
-        await event.message.respond(f"{event.author.mention} I do work.")
-    elif findWholeWord('hey', messageContent) or findWholeWord('hi', messageContent) or findWholeWord('hello', messageContent):
-        await event.message.respond(f"Hey {event.author.mention}, I am a cat. With robot intestines. If you're bored, you should ask me a question, or check out my `+userinfo`, `+ping`, `+fortune` and `+fact` commands :cat:")
-    else:
-        await event.message.respond(f"{event.author.mention}, did you forget a question mark? <:mmhmmm:872809423939174440>")
 
 """
 Bot adds #notalurker role for those who comment.
